DefaultPrecision now also changes that for complex numbers.  also, a refactor on random.
