+++ core/include/bertini2/mpfr_complex.hpp
@@ -34,7 +34,7 @@
 
 #include "bertini2/config.h"
 
-#include "bertini2/mpfr_extensions.hpp"
+
 
 #include <boost/archive/text_oarchive.hpp>
 #include <boost/archive/text_iarchive.hpp>
@@ -45,124 +45,21 @@
 
 #include <boost/multiprecision/mpc.hpp>
 
+#include "bertini2/mpfr_extensions.hpp"
 
 namespace bertini{
-namespace multiprecision{
+
 
 namespace bmp = boost::multiprecision;
 using bmp::backends::mpc_complex_backend;
-using complex = bmp::number<mpc_complex_backend<0>, bmp::et_on >;
 
-/** 
-\brief Get the precision of a number.
+#ifdef BMP_EXPRESSION_TEMPLATES
+	using mpfr_complex = bmp::number<mpc_complex_backend<0>, bmp::et_on >;
+#else
+	using mpfr_complex = bmp::number<mpc_complex_backend<0>, bmp::et_off >;
+#endif
 
-For bertini::mpfr_complex, this calls the precision member method for bertini::mpfr_complex.
-*/
-inline
-unsigned Precision(complex const& num)
-{
-	return num.precision();
-}
-
-inline void Precision(complex & num, unsigned prec)
-{
-	num.precision(prec);
-}
-
-// inline 
-// bool isnan(complex const& num)
-// {
-// 	return num.isnan();
-// }
-
-inline 
-void RandomReal(complex & a, unsigned num_digits)
-{
-	a.precision(num_digits);
-	RandomMp(a.real(),num_digits);
-	a.imag() = 0;
-}
-
-inline 
-void rand(complex & a, unsigned num_digits)
-{
-	a.precision(num_digits);
-	RandomMp(a.real(),num_digits);
-	RandomMp(a.imag(),num_digits);
-}
-
-inline 
-void RandomComplex(complex & a, unsigned num_digits)
-{
-	rand(a,num_digits);
-}
-
-
-		/**
-		 Produce a random complex number, to default precision.
-		 */
-		inline complex rand()
-		{
-			complex returnme( RandomMp(mpfr_float(-1),mpfr_float(1)), RandomMp(mpfr_float(-1),mpfr_float(1)) );
-			returnme /= sqrt( abs(returnme));
-			return returnme;
-		}
-		
-		inline complex RandomUnit()
-		{
-			complex returnme( RandomMp(mpfr_float(-1),mpfr_float(1)), RandomMp(mpfr_float(-1),mpfr_float(1)) );
-			returnme /= abs(returnme);
-			return returnme;
-		}
-		/**
-		 Produce a random real number \f$\in [-1,\,1]\f$, to current default precision. 
-		 */
-		inline complex RandomReal()
-		{
-			using std::sqrt;
-			complex returnme( RandomMp(mpfr_float(-1),mpfr_float(1)), RandomMp(mpfr_float(-1),mpfr_float(1)) );
-			returnme /= sqrt( abs(returnme));
-			return returnme;
-		}
-		
-
-
-inline 
-complex RandomComplex(unsigned num_digits)
-{
-	complex z;
-	RandomComplex(z, num_digits);
-	return z;
-}
-
-inline 
-void RandomUnit(complex & a, unsigned num_digits)
-{
-	auto prev_precision = DefaultPrecision();
-
-	a.precision(num_digits);
-	RandomMp(a.real(),num_digits);
-	RandomMp(a.imag(),num_digits);
-	a /= abs(a);
-
-	DefaultPrecision(prev_precision);
-}
-
-inline 
-complex RandomUnit(unsigned num_digits)
-{
-	complex a;
-	RandomUnit(a,num_digits);
-	return a;
-}
-
-
-
-
-
-}  // namespace multiprecision
-
-using mpfr_complex = multiprecision::complex;
+	using dbl = std::complex<double>;
 
 	/** 
 	\brief Get the precision of a number.
@@ -189,8 +86,25 @@ using mpfr_complex = multiprecision::complex;
 	inline
 	bool isnan(bertini::mpfr_complex const& num){return isnan(num.real()) || isnan(num.imag());};
 
-} // namespaces
-// }
+
+
+
+	inline auto DefaultPrecision()
+	{
+		auto p = mpfr_float::default_precision();
+		assert(p==mpfr_complex::default_precision() && "precision of real and complex multiprecision numbers have drifted...");
+		return p;
+	}
+
+	inline void DefaultPrecision(unsigned prec)
+	{
+		mpfr_float::default_precision(prec);
+		mpfr_complex::default_precision(prec);
+	}
+}
+
+
+
